backend-files/park_on_go_backend-master/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-environ==0.11.2
django-rest==0.8.7
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.4.4
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
